@@ -5,41 +5,44 @@
 //  Created by 박병관 on 6/5/24.
 //
 
-import XCTest
+import Testing
 #if canImport(CoreData)
 import CoreData
 @testable import Tetra
+internal import NamespaceExtension
 
-final class NSManagedObjectContextTests: XCTestCase {
+@Suite
+struct NSManagedObjectContextTests {
 
 
+    @Test
     func testBlocking() throws {
         let context = NSManagedObjectContext(concurrencyType: .privateQueueConcurrencyType)
         let uuid = UUID()
         let returnValue = context.tetra._performAndWait {
             return uuid
         }
-        XCTAssertEqual(uuid, returnValue)
+        #expect(uuid == returnValue)
         let result = Result {
             try context.tetra._performAndWait {
                 throw CancellationError()
             }
         }
-        XCTAssertThrowsError(try result.get()) {
-            XCTAssertTrue($0 is CancellationError, "\($0) is not \(CancellationError.self)")
-
-        }
+        #expect(throws: CancellationError.self, performing: {
+            try result.get()
+        })
         
     }
     
+    @Test
     func testMainAsync() async throws {
         let context = NSManagedObjectContext(concurrencyType: .mainQueueConcurrencyType)
         let uuid = UUID()
         let returnValue = await context.tetra._performEnqueue {
-            XCTAssertTrue(Thread.isMainThread)
+            #expect(Thread.isMainThread)
             return uuid
         }
-        XCTAssertEqual(uuid, returnValue)
+        #expect(uuid == returnValue)
         let result:Result<Void,any Error>
         do {
             try await context.tetra._performEnqueue{
@@ -49,19 +52,20 @@ final class NSManagedObjectContextTests: XCTestCase {
         } catch {
             result = .failure(error)
         }
-        XCTAssertThrowsError(try result.get()) {
-            XCTAssertTrue($0 is CancellationError, "\($0) is not \(CancellationError.self)")
-        }
+        #expect(throws: CancellationError.self, performing: {
+            try result.get()
+        })
     }
     
+    @Test
     func testBackgroundAsync() async throws {
         let context = NSManagedObjectContext(concurrencyType: .privateQueueConcurrencyType)
         let uuid = UUID()
         let returnValue = await context.tetra._performEnqueue {
-            XCTAssertFalse(Thread.isMainThread)
+            #expect(!Thread.isMainThread)
             return uuid
         }
-        XCTAssertEqual(uuid, returnValue)
+        #expect(uuid == returnValue)
         let result:Result<Void,any Error>
         do {
             try await context.tetra._performEnqueue{
@@ -71,41 +75,39 @@ final class NSManagedObjectContextTests: XCTestCase {
         } catch {
             result = .failure(error)
         }
-        XCTAssertThrowsError(try result.get()) {
-            XCTAssertTrue($0 is CancellationError, "\($0) is not \(CancellationError.self)")
+        #expect(throws: CancellationError.self) {
+            try result.get()
         }
     }
     
-    
-    func testImmediate() throws {
+    @Test
+    func testImmediate() async throws {
         let context = NSManagedObjectContext(concurrencyType: .privateQueueConcurrencyType)
-        let completion = expectation(description: "completion")
-        context.perform {
-            defer { completion.fulfill() }
-            let uuid = UUID()
-            let expectValue = context.tetra._performImmediate {
-                return uuid
-            }
-            XCTAssertEqual(expectValue, .success(uuid))
-            let expectThrow = Result {
-                try context.tetra._performImmediate {
-                    throw CancellationError()
+        let _:Void = await withUnsafeContinuation{ continuation in
+            context.perform {
+                defer {
+                    continuation.resume()
+                }
+                let uuid = UUID()
+                let expectValue = context.tetra._performImmediate {
+                    return uuid
+                }
+                #expect(expectValue == .success(uuid))
+                let expectThrow = Result {
+                    try context.tetra._performImmediate {
+                        throw CancellationError()
+                    }
+                }
+                #expect(throws: CancellationError.self) {
+                    try expectThrow.get()
                 }
             }
-            func checkError() {
-                XCTAssertThrowsError(try expectThrow.get()) {
-                    XCTAssertTrue($0 is CancellationError, "\($0) is not \(CancellationError.self)")
-                }
-            }
-            checkError()
-        }
-        
-        let expectNil = context.tetra._performImmediate {
             
+            let expectNil = context.tetra._performImmediate {
+                
+            }
+            #expect(expectNil == nil)
         }
-        XCTAssertNil(expectNil)
-        
-        wait(for: [completion], timeout: 0.1)
         
     }
     
