@@ -8,8 +8,9 @@
 import Foundation
 import Dispatch
 import Combine
+internal import CriticalSection
+import Namespace
 
-extension DispatchSource: TetraExtended {}
 
 public extension TetraExtension where Base:DispatchSource {
     
