@@ -9,77 +9,68 @@ import Foundation
 import _Concurrency
 
 #if canImport(CoreData)
-import CoreData
+@preconcurrency import CoreData
+import Namespace
 
-extension NSPersistentContainer: TetraExtended {}
-extension NSPersistentStoreCoordinator: TetraExtended {}
-extension NSManagedObjectContext: TetraExtended {}
+
 
 extension TetraExtension where Base: NSPersistentStoreCoordinator {
     
     @usableFromInline
-    internal func _perform<T>(_ body: () throws -> T) async rethrows -> T {
-        let result:Result<T,Error>
-        do {
-            let value: T = try await withoutActuallyEscaping(body) { escapingClosure in
-                let closureHolder = ClosureHolder(closure: escapingClosure)
-                defer {
-                    withExtendedLifetime(closureHolder) {}
-                }
-                return try await withUnsafeThrowingContinuation { continuation in
-                    base.perform{ [unowned closureHolder, continuation] in
-                        continuation.resume(with: Result { try closureHolder.closure() })
-                    }
+    internal func _perform<T,Failure:Error>(_ body: () throws(Failure) -> T) async throws(Failure) -> T {
+        let value:Result<T,Failure> = await withoutActuallyEscaping(body) { escapingClosure in
+            let block = ClosureHolder(closure: escapingClosure)
+            defer {
+                withExtendedLifetime(block, {})
+            }
+            return await withUnsafeContinuation { continuation in
+                base.perform { [unowned block, continuation] in
+                    continuation.resume(returning: block())
                 }
             }
-            result = .success(value)
-        } catch {
-            result = .failure(error)
         }
-        switch result {
-        case .success(let success):
-            return success
-        case .failure:
-            try result._rethrowOrFail()
-        }
+        return try value.get()
     }
     
     @inlinable
-    public func perform<T>(_ body: () throws -> T) async rethrows -> T {
+    public func perform<T,Failure:Error>(_ body: () throws(Failure) -> T) async throws(Failure) -> T {
         return if #available(iOS 15.0, tvOS 15.0, macCatalyst 15.0, watchOS 8.0, macOS 12.0, *) {
-            try await withoutActuallyEscaping(body) {
-                let holder = ClosureHolder(closure: $0)
-                defer { withExtendedLifetime(holder, {})}
-                return try await base.perform{ [unowned holder] in
-                    try holder.closure()
+            try await withoutActuallyEscaping(body) { 
+                let block = ClosureHolder(closure: $0)
+                defer {
+                    withExtendedLifetime(block, {})
                 }
-            }
+                return await base.perform{ [unowned block] in
+                    block()
+                }
+            }.get()
         } else {
             try await _perform(body)
         }
     }
     
     @usableFromInline
-    internal func _performAndWait<T>(_ body: () throws -> T) rethrows -> T {
-        var result:Result<T,any Error>? = nil
+    internal func _performAndWait<T,Failure:Error>(_ body: () throws(Failure) -> T) throws(Failure) -> T {
+        var result:Result<T,Failure>? = nil
         base.performAndWait {
-            result = Result { try body() }
+            do throws(Failure) {
+                result = .success(try body())
+            } catch {
+                result = .failure(error)
+            }
         }
         guard let result else {
             preconditionFailure("performAndWait didn't run")
         }
-        switch result {
-        case .success(let success):
-            return success
-        case .failure:
-            try result._rethrowOrFail()
-        }
+        return try result.get()
     }
     
     @inlinable
-    public func performAndWait<T>(_ body: () throws -> T) rethrows -> T {
+    public func performAndWait<T,Failure:Error>(_ body: () throws(Failure) -> T) throws(Failure) -> T {
         return if #available(iOS 15.0, tvOS 15.0, macCatalyst 15.0, watchOS 8.0, macOS 12.0, *) {
-            try base.performAndWait(body)
+            try base.performAndWait{
+                Result(catching: body)
+            }.get()
         } else {
             try _performAndWait(body)
         }
@@ -104,110 +95,104 @@ extension TetraExtension where Base: NSManagedObjectContext {
     ///
     /// This method supports reentrancy — meaning it’s safe to call the method again, from within the closure, before the previous invocation completes.
     @usableFromInline
-    internal func _performImmediate<T>(
-        _ body: () throws -> T
-    ) rethrows -> Result<T,Never>? {
+    internal func _performImmediate<T,Failure:Error>(
+        _ body: () throws(Failure) -> T
+    ) throws(Failure) -> Result<T,Never>? {
         // Suppress deprecation
         let lock:any ObjcLocking = base
         // NSManagedObjectContext has Reentrant Locking
         guard lock.tryLock() else { return nil }
         defer { lock.unlock() }
-        let value = try performAndWait(body)
+        let value = try _performAndWait(body)
         return .success(value)
     }
     
     @usableFromInline
-    internal func _performEnqueue<T>(
-        _ body: () throws -> T
-    ) async rethrows -> T {
-        let result:Result<T,any Error>
-        do {
-            let value: T = try await withoutActuallyEscaping(body) { escapingClosure in
-                let closureHolder = ClosureHolder(closure: escapingClosure)
-                defer {
-                    withExtendedLifetime(closureHolder) {}
-                }
-                return try await withUnsafeThrowingContinuation { continuation in
-                    base.perform{ [unowned closureHolder, continuation] in
-                        continuation.resume(with: Result { try closureHolder.closure() })
-                    }
+    internal func _performEnqueue<T,Failure:Error>(
+        _ body: () throws(Failure) -> T
+    ) async throws(Failure) -> T {
+        let result: Result<T,Failure> = await withoutActuallyEscaping(body) { escapingClosure in
+            let holder = ClosureHolder(closure: escapingClosure)
+            defer {
+                withExtendedLifetime(holder, {})
+            }
+            return await withUnsafeContinuation { continuation in
+                base.perform{ [unowned(unsafe) holder, continuation] in
+                    let result = holder()
+                    continuation.resume(returning: result)
                 }
             }
-            result = .success(value)
-        } catch {
-            result = .failure(error)
         }
-        switch result {
-        case .success(let success):
-            return success
-        case .failure:
-            try result._rethrowOrFail()
-        }
-    }
-    
-    /// Asynchronously performs the specified closure on the context’s queue.
-    @inlinable
-    @_unsafeInheritExecutor
-    public func perform<T>(
-        schedule: CoreDataScheduledTaskType = .immediate,
-        _ body: () throws -> T
-    ) async rethrows -> T {
-        /*
-         
-         Since this method and NSManagedObjectContext peform has no actor preference and isolation restriction.
-         These two are always called on global nonisolated context (Actor switching happen).
-         Which means that `immediate` execution option is totally no-op.
-         
-         
-         - `_performImmediate:` is never called when using `NSManagedObjectContext.perform(schedule: .immediate)` in iOS 15 ~ iOS 17
-         
-         - @_unsafeInheritExecutor do fix the above problem
-         */
-        return if #available(iOS 15.0, tvOS 15.0, macCatalyst 15.0, watchOS 8.0, macOS 12.0, *) {
-            try await withoutActuallyEscaping(body) {
-                let holder = ClosureHolder(closure: $0)
-                defer {
-                    withExtendedLifetime(holder, { })
-                }
-                return try await base.perform(schedule: schedule.platformValue) { [unowned holder] in
-                    try holder.closure()
-                }
-            }
-        } else if schedule == .enqueued {
-            try await _performEnqueue(body)
-        } else {
-            if let result = try _performImmediate(body) {
-                switch result {
-                case .success(let success):
-                    success
-                }
-            } else {
-                try await _performEnqueue(body)
-            }
-        }
+        return try result.get()
     }
     
     @usableFromInline
-    internal func _performAndWait<T>(_ body: () throws -> T) rethrows -> T {
-        var result:Result<T,any Error>? = nil
+    internal func _perform<T, Failure:Error>(
+        schedule:CoreDataScheduledTaskType = .immediate,
+        isolation: isolated (any Actor)? = #isolation,
+        _ body: @escaping () throws(Failure) -> T
+    ) async throws(Failure) -> Suppress<T> {
+        nonisolated(unsafe)
+        let callRef = self
+        let ref = UnsafeClosureHolder(closure: body)
+        defer {
+            withExtendedLifetime(ref, {})
+        }
+        if schedule == .immediate, let result = try _performImmediate(body) {
+            let value = result.get()
+            return .init(value: value)
+        } else {
+            return try await callRef._performEnqueue { [unowned(unsafe) ref] in
+                ref().map(Suppress.init)
+            }.get()
+        }
+    }
+    
+    
+    /// Asynchronously performs the specified closure on the context’s queue.
+    @inlinable
+    public func perform<T, Failure:Error>(
+        schedule:CoreDataScheduledTaskType = .immediate,
+        isolation: isolated (any Actor)? = #isolation,
+        _ body: () throws(Failure) -> T
+    ) async throws(Failure) -> T {
+        let box: Suppress<T> = try await withoutActuallyEscaping(body) { escapingClosure async throws(Failure) in
+            if #available(iOS 15.0, tvOS 15.0, macCatalyst 15.0, watchOS 8.0, macOS 12.0, *) {
+                let ref:UnsafeClosureHolder<T,Failure> = UnsafeClosureHolder(closure: escapingClosure)
+                defer {
+                    withExtendedLifetime(ref, {})
+                }
+                let wrapped = await base.perform(schedule: schedule.platformValue) { [unowned(unsafe) ref] in
+                    ref().map(Suppress.init)
+                }
+                return try wrapped.get()
+            } else {
+                return try await self._perform(schedule: schedule, isolation: isolation, escapingClosure)
+            }
+        }
+        return box.value
+    }
+    
+    @usableFromInline
+    internal func _performAndWait<T,Failure:Error>(_ body: () throws(Failure) -> T) throws(Failure) -> T {
+        var result:Result<T,Failure>? = nil
         base.performAndWait {
-            result = Result { try body() }
+            do throws(Failure) {
+                result = .success(try body())
+            } catch {
+                result = .failure(error)
+            }
         }
         guard let result else {
             preconditionFailure("performAndWait didn't run")
         }
-        switch result {
-        case .success(let success):
-            return success
-        case .failure:
-            try result._rethrowOrFail()
-        }
+        return try result.get()
     }
     
     @inlinable
-    public func performAndWait<T>(_ body: () throws -> T) rethrows -> T {
+    public func performAndWait<T,Failure:Error>(_ body: () throws(Failure) -> T) throws(Failure) -> T {
         return if #available(iOS 15.0, tvOS 15.0, macCatalyst 15.0, watchOS 8.0, macOS 12.0, *) {
-            try base.performAndWait(body)
+            try base.performAndWait{ Result(catching: body) }.get()
         } else {
             try _performAndWait(body)
         }
@@ -218,50 +203,42 @@ extension TetraExtension where Base: NSManagedObjectContext {
 extension TetraExtension where Base: NSPersistentContainer {
     
     @inlinable
-    public func performBackground<T>(_ body: (NSManagedObjectContext) throws -> T) async rethrows -> T {
+    public func performBackground<T,Failure:Error>(_ body: (NSManagedObjectContext) throws(Failure) -> T) async throws(Failure) -> T {
         return if #available(iOS 15.0, tvOS 15.0, macCatalyst 15.0, watchOS 8.0, macOS 12.0, *) {
             try await withoutActuallyEscaping(body) {
-                let holder = CoreDataContextClosureHolder(closure: $0)
-                defer {
-                    withExtendedLifetime(holder, { })
+                let block = CoreDataContextClosureHolder(closure: $0)
+                defer { withExtendedLifetime(block, {}) }
+                return await base.performBackgroundTask{ [unowned block] in
+                    block($0)
                 }
-                return try await base.performBackgroundTask{ [unowned holder] in
-                    try holder.closure($0)
-                }
-            }
+            }.get()
         } else {
             try await _performBackground(body)
         }
     }
     
+
+    
     @usableFromInline
-    internal func _performBackground<T>(_ body: (NSManagedObjectContext) throws -> T) async rethrows -> T {
-        let result:Result<T,Error>
-        do {
-            let value: T = try await withoutActuallyEscaping(body) { escapingClosure in
-                let closureHolder = CoreDataContextClosureHolder(closure: escapingClosure)
-                defer {
-                    withExtendedLifetime(closureHolder) {}
-                }
-                return try await withUnsafeThrowingContinuation { continuation in
-                    base.performBackgroundTask { [unowned closureHolder, continuation] newContext in
-                        continuation.resume(with: Result{ try closureHolder.closure(newContext) })
-                    }
+    internal func _performBackground<T,Failure:Error>(_ body: (NSManagedObjectContext) throws(Failure) -> T) async throws(Failure) -> T {
+        let result:Result<T,Failure> = await withoutActuallyEscaping(body) { escapingClosure in
+            let holder = CoreDataContextClosureHolder(closure: escapingClosure)
+            defer {
+                withExtendedLifetime(holder, {})
+            }
+            return await withUnsafeContinuation { continuation in
+                base.performBackgroundTask { [unowned holder, continuation] newContext in
+                    nonisolated(unsafe)
+                    let result = holder(newContext)
+                    continuation.resume(returning: result)
                 }
             }
-            result = .success(value)
-        } catch {
-            result = .failure(error)
         }
-        switch result {
-        case .success(let success):
-            return success
-        case .failure:
-            try result._rethrowOrFail()
-        }
+        return try result.get()
     }
     
 }
+
 
 public enum CoreDataScheduledTaskType: Sendable, Hashable {
     
@@ -282,3 +259,31 @@ public enum CoreDataScheduledTaskType: Sendable, Hashable {
 }
 
 #endif
+
+extension TetraExtension: Sendable where Base: Sendable {}
+
+@usableFromInline
+internal final class UnsafeClosureHolder<R:~Copyable,Failure:Error>: @unchecked Sendable {
+    
+    @inline(__always)
+    @usableFromInline let closure: () throws(Failure) -> R
+    
+    @inline(__always)
+    @inlinable
+    init(closure: @escaping  () throws(Failure) -> R) {
+        self.closure = closure
+    }
+    
+    @inline(__always)
+    @inlinable
+    func callAsFunction() -> Result<R,Failure> {
+        do {
+            let value = try closure()
+            return .success(value)
+        } catch {
+            return .failure(error)
+        }
+    }
+    
+}
+
