@@ -4,62 +4,51 @@
 //
 //  Created by pbk on 2023/01/27.
 //
-
-import XCTest
+import Foundation
 @testable import Tetra
+import Testing
 
-final class AnyEncodableTests: XCTestCase {
+@Suite
+struct AnyEncodableTests {
 
+    @Test
     func testURLEncoding() throws {
         
         let targetURL = FileManager.default.temporaryDirectory
-        
-        XCTAssertEqual(
-            try JSONEncoder().encode(AnyEncodable(targetURL)),
-            try JSONEncoder().encode(targetURL)
-        )
-        
-        try XCTExpectFailure {
-            XCTAssertEqual(
-                try JSONEncoder().encode(AnyErasedEncodable(value: targetURL)),
-                try JSONEncoder().encode(targetURL)
-            )
-        }
-
-
+        let defaultValue =  try JSONEncoder().encode(targetURL)
+        let anyValue = try JSONEncoder().encode(AnyEncodable(targetURL))
+        let erasedValue = try JSONEncoder().encode(AnyErasedEncodable(value: targetURL))
+        #expect(defaultValue == anyValue)
+        #expect(defaultValue != erasedValue)
     }
-
     
+
+    @Test
     func testURLObjectEncoding() throws {
         let targetURL = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
         let objectForm = ["A": targetURL, "B": targetURL]
         let wrappedFrom = AnyEncodable(objectForm)
-        XCTAssertEqual(
-            try JSONSerialization.jsonObject(with: JSONEncoder().encode(objectForm)) as! NSDictionary,
-            try JSONSerialization.jsonObject(with: JSONEncoder().encode(wrappedFrom)) as! NSDictionary
-        )
-        
-        XCTAssertEqual(
-            try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(objectForm), format: nil) as! NSDictionary,
-            try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(wrappedFrom), format: nil) as! NSDictionary
-        )
+        var wrappedObject = try JSONSerialization.jsonObject(with: JSONEncoder().encode(wrappedFrom)) as! NSDictionary
+        var rawObject = try JSONSerialization.jsonObject(with: JSONEncoder().encode(objectForm)) as! NSDictionary
+        #expect(wrappedObject == rawObject)
+        rawObject = try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(objectForm), format: nil) as! NSDictionary
+        wrappedObject = try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(wrappedFrom), format: nil) as! NSDictionary
+        #expect(rawObject == wrappedObject)
     }
     
+    @Test
     func testURLArrayEncoding() throws {
         let targetURL = FileManager.default.temporaryDirectory
         let arrayForm = (0..<10).map{ _ in
             targetURL.appendingPathComponent(UUID().uuidString)
         }
         let wrappedFrom = AnyEncodable(arrayForm)
-        XCTAssertEqual(
-            try JSONSerialization.jsonObject(with: JSONEncoder().encode(arrayForm)) as! NSArray,
-            try JSONSerialization.jsonObject(with: JSONEncoder().encode(wrappedFrom)) as! NSArray
-        )
-        
-        XCTAssertEqual(
-            try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(arrayForm), format: nil) as! NSArray,
-            try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(wrappedFrom), format: nil) as! NSArray
-        )
+        var rawArray = try JSONSerialization.jsonObject(with: JSONEncoder().encode(arrayForm)) as! NSArray
+        var wrappedArray = try JSONSerialization.jsonObject(with: JSONEncoder().encode(wrappedFrom)) as! NSArray
+        #expect(rawArray == wrappedArray)
+        rawArray = try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(arrayForm), format: nil) as! NSArray
+        wrappedArray = try PropertyListSerialization.propertyList(from: PropertyListEncoder().encode(wrappedFrom), format: nil) as! NSArray
+        #expect(rawArray == wrappedArray)
     }
     
 }
