@@ -9,6 +9,8 @@ import Foundation
 import XCTest
 @testable import Tetra
 import Combine
+import BackPortAsyncSequence
+import Namespace
 
 class AsyncSequencePublisherTests: XCTestCase {
     
@@ -22,7 +24,8 @@ class AsyncSequencePublisherTests: XCTestCase {
             source.forEach{ continuation.yield($0) }
             continuation.finish()
         }
-        let cancellable = AsyncSequencePublisher(base: stream)
+        let cancellable = AsyncTypedStream(base: stream)
+            .tetra.toPublisher()
             .catch{ _ in
                 XCTFail()
                 return Empty<Int, Never>()
@@ -59,7 +62,7 @@ class AsyncSequencePublisherTests: XCTestCase {
             }
             return value
         }
-        let pub = AsyncSequencePublisher(base: asyncSequence)
+        let pub = AsyncSequencePublisher(base: LegacyTypedAsyncSequence(base: asyncSequence))
             .handleEvents(
                 receiveCancel: { expect.fulfill() }
             )
@@ -88,7 +91,7 @@ class AsyncSequencePublisherTests: XCTestCase {
             }
             return value
         }
-        let cancellable = AsyncSequencePublisher(base: asyncSequence)
+        let cancellable = AsyncSequencePublisher(base: LegacyTypedAsyncSequence(base: asyncSequence))
             .mapError{
                 $0 as! CancellationError
             }
